+++ cmd/nightshift/commands/root.go
@@ -10,7 +10,7 @@ import (
 
 var (
 	// Version is set at build time
-	Version = "0.3.4"
+	Version = "1.0.0"
 )
 
 var rootCmd = &cobra.Command{